refactor: move utility functions into the utils folder
